charge_service/requirements.txt
pydantic-settings
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
beanie==1.22.6
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
idna==3.4
iniconfig==2.0.0
lazy-model==0.2.0
mongomock==4.1.2
mongomock-motor==0.0.21
motor==3.3.1
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pydantic_core==2.10.1
pymongo==4.5.0
pytest==7.4.2
pytest-mock==3.11.1
sentinels==1.0.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
aio-pika
aiosqlite
fastapi-scheduler
mongoengine
redis
pymongo
pydantic[dotenv]
python-decouple
fastapi==0.103.2
pydantic[email]
uvicorn==0.17.4
PyJWT
bcrypt
passlib
locust
pytest
httpx
orjson==3.9.10 
celery==5.3.6
flower
kombu==5.3.5